Added choice of SSM knockouts Fixed evaluation script to perform SSM knockout Fixed eval script to continue from checkpoint

+++ scripts/evaluate_context_interference.py
@@ -13,6 +13,8 @@ from src.knockout.ssm_knockout import SSMKnockoutEvaluator
 from argparse import ArgumentParser, Namespace
 import numpy as np
 from src.utils import load_knowns, setup_model
+from src.knockout.ssm_knockout.ssm_classifier import SSMClassifier, DecayNormClassifier
+from typing import Optional
 
 
 def get_args():
@@ -22,6 +24,8 @@ def get_args():
     parser.add_argument("--drop_subj_last", action='store_true')
     parser.add_argument("--show_eval_progress", action='store_true')
     parser.add_argument("--output_dir", type=Path, default=Path("resources"))
+    parser.add_argument("--layer_checkpoint", type=Path, default=None)
+    parser.add_argument("--bin_search_checkpoint", type=Path, default=None)
     # parser.add_argument("--affected_outputs", type=str, default="LAST", options=['LAST', 'ENTIRE_SUBJ'])
     return parser.parse_args()
 
@@ -109,11 +113,11 @@ def get_last_token_stats(model_size: str = '130M'):
     print(stat)
 
 
-def attention_knockout_evaluate(args: Namespace, model: MambaForCausalLM, tokenizer: AutoTokenizer, device: torch.device, knowns_df: pd.DataFrame):
+def attention_knockout_evaluate(args: Namespace, model: MambaForCausalLM, tokenizer: AutoTokenizer, device: torch.device, knowns_df: pd.DataFrame, layer_checkpoint: Optional[pd.DataFrame] = None, bin_search_checkpoint: Optional[pd.DataFrame] = None):
     evaluator = AttentionKnockoutEvaluator(model, tokenizer, device, -1, -1, args.drop_subj_last, args.show_eval_progress)
 
-    bin_search_df = None
-    layer_df = None
+    bin_search_df = bin_search_checkpoint
+    layer_df = layer_checkpoint
 
     affected_outputs = [KnockoutTarget.LAST, KnockoutTarget.ENTIRE_SUBJ]
 
@@ -121,32 +125,35 @@ def attention_knockout_evaluate(args: Namespace, model: MambaForCausalLM, tokeni
     
     for output in affected_outputs:
         for target in specific_targets[output]:
+            
             evaluator.knockout_target = target
             evaluator.affected_target = output
             
-            curr_df = binary_search(evaluator, knowns_df, KnockoutMode[args.interfere_mode])
-            curr_df['knockout_inputs'] = target
-            curr_df['affected_outputs'] = output
-            bin_search_df = [bin_search_df, curr_df]
-            bin_search_df = pd.concat(bin_search_df)
+            if not ((target in bin_search_df['knockout_inputs'].unique()) & (output in bin_search_df['affected_outputs'].unique())):
+                curr_df = binary_search(evaluator, knowns_df, KnockoutMode[args.interfere_mode])
+                curr_df['knockout_inputs'] = target
+                curr_df['affected_outputs'] = output
+                bin_search_df = [bin_search_df, curr_df]
+                bin_search_df = pd.concat(bin_search_df)
             
-            # save to csv
-            out_fname = args.output_dir / f"{args.interfere_mode}_{args.model_size}_bin_search.csv"
-            if out_fname.exists():
-                os.remove(out_fname)
-            bin_search_df.to_csv(out_fname)
+                # save to csv
+                out_fname = args.output_dir / f"{args.interfere_mode}_{args.model_size}_bin_search.csv"
+                if out_fname.exists():
+                    os.remove(out_fname)
+                bin_search_df.to_csv(out_fname)
             
-            curr_df = layer_by_layer(evaluator, knowns_df, KnockoutMode[args.interfere_mode])
-            curr_df['knockout_inputs'] = target
-            curr_df['affected_outputs'] = output
-            layer_df = [layer_df, curr_df]
-            layer_df = pd.concat(layer_df)
-            
-            # save to csv
-            out_fname = args.output_dir / f"{args.interfere_mode}_{args.model_size}_layer_by_layer.csv"
-            if out_fname.exists():
-                os.remove(out_fname)
-            layer_df.to_csv(out_fname)
+            if not ((target in layer_df['knockout_inputs'].unique()) & (output in layer_df['affected_outputs'].unique())):
+                curr_df = layer_by_layer(evaluator, knowns_df, KnockoutMode[args.interfere_mode])
+                curr_df['knockout_inputs'] = target
+                curr_df['affected_outputs'] = output
+                layer_df = [layer_df, curr_df]
+                layer_df = pd.concat(layer_df)
+                
+                # save to csv
+                out_fname = args.output_dir / f"{args.interfere_mode}_{args.model_size}_layer_by_layer.csv"
+                if out_fname.exists():
+                    os.remove(out_fname)
+                layer_df.to_csv(out_fname)
 
 
 def layer_knockout_evaluate(args: Namespace, model: MambaForCausalLM, tokenizer: AutoTokenizer, device: torch.device, knowns_df: pd.DataFrame):
@@ -159,12 +166,15 @@ def layer_knockout_evaluate(args: Namespace, model: MambaForCausalLM, tokenizer:
 
 
 def ssm_knockout_evaluate(args: Namespace, model: MambaForCausalLM, tokenizer: AutoTokenizer, device: torch.device, knowns_df: pd.DataFrame):
+    ssm_classifier = DecayNormClassifier()
     bin_search_df = None
     layer_df = None
-    for norm in ['l_1','l_infty']:
-        categorized_As = {'category_A':[[1,2,3]] * 24, 'category_B': [[4,5,6]]  * 24}
-        for category in categorized_As:
-            evaluator = SSMKnockoutEvaluator(model, tokenizer, device, categorized_As[category], False)
+    for norm in [1, float('inf')]:
+        ssm_classifier.norm = norm
+
+        categorized_ssms = ssm_classifier.classify_model(model.backbone)
+        for category in categorized_ssms:
+            evaluator = SSMKnockoutEvaluator(model, tokenizer, device, categorized_ssms[category], False)
             curr = binary_search(evaluator, knowns_df, KnockoutMode[args.interfere_mode])
             curr['category'] = category
             curr['norm'] = norm
@@ -189,6 +199,12 @@ def ssm_knockout_evaluate(args: Namespace, model: MambaForCausalLM, tokenizer: A
     layer_df.to_csv(args.output_dir / f"{args.interfere_mode}_{args.model_size}_layer_by_layer.csv")
 
 
+def get_checkpoint(pth: Optional[Path]) -> Optional[pd.DataFrame]:
+    if pth is not None:
+        return pd.read_csv(pth)
+    return None
+
+
 def main() -> None:
     args = get_args()
     get_last_token_stats(args.model_size)
@@ -196,17 +212,21 @@ def main() -> None:
     model, tokenizer, device = setup_model(args.model_size)
     knowns_df = load_knowns()
     args.output_dir.mkdir(parents=True, exist_ok=True)
+
+    bin_search_checkpoint = get_checkpoint(args.bin_search_checkpoint)
+    layer_checkpoint = get_checkpoint(args.layer_checkpoint)
     
     # If we do attention knockout:
     if KnockoutMode[args.interfere_mode] in {KnockoutMode.ZERO_ATTENTION, KnockoutMode.ZERO_DELTA}:
-        attention_knockout_evaluate(args, model, tokenizer, device, knowns_df)
-        
+        attention_knockout_evaluate(args, model, tokenizer, device, knowns_df, layer_checkpoint, bin_search_checkpoint)
+
     # If we skip entire layer \ component
     elif KnockoutMode[args.interfere_mode] in {KnockoutMode.IGNORE_CONTEXT, KnockoutMode.IGNORE_LAYER, KnockoutMode.ONLY_CONTEXT}:
         layer_knockout_evaluate(args, model, tokenizer, device, knowns_df)
+
+    # If we do SSM knockout
     elif KnockoutMode[args.interfere_mode] == KnockoutMode.IGNORE_SSM:
         ssm_knockout_evaluate(args, model, tokenizer, device, knowns_df)
-        
     else:
         raise ValueError(f"Unknown knockout mode: {args.interfere_mode}")
